Enable automated fixes in Qodana linter
The changes in the code make the Qodana linter automatic fixes available in the GitHub workflow. The fixesStrategy has been commented out in the qodana.yaml file and the --apply-fixes argument was added in the code-quality.yml, implying that it will try to auto-fix the issues detected during the linting process. The updated workflow will also push these fixes directly to the pull request.